requirements.txt
fastapi
uvicorn[standard]
jinja2
pyserial
